BUG: COPY_FROM is more correct than MOVE_FROM.
When a patch is generated for a directory which is partially copied and
partially moved, the Differ uses MOVE_FROM which does not work
correctly. This commit adds a test case which demonstrates the problem
with a simple set of files. The commit also fixes the bug in a
simplistic manner that passes all tests.